requirements.txt
# Atlas-Podemos Unified Requirements
# Merged dependencies from both Atlas cognitive amplification platform 
# and Podemos advanced podcast processor

# Core Framework Dependencies
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
beautifulsoup4==4.13.4
black
certifi==2025.7.14
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cssselect==1.3.0
defusedxml==0.7.1
distro==1.9.0

# Web Framework & API
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.1
Jinja2==3.1.6
python-multipart==0.0.20

# RSS Feed Processing (Podemos)
feedparser==6.0.11

# Database & Storage
SQLAlchemy==2.0.41
greenlet==3.2.3

# Configuration & Environment
python-dotenv==1.1.1
PyYAML==6.0.2
pydantic==2.11.7
pydantic_core==2.33.2

# HTTP & Networking
httpcore==1.0.9
httpx==0.28.1
requests==2.32.4
urllib3==2.5.0
idna==3.10

# Audio Processing & Transcription (Podemos Enhanced)
faster-whisper
filelock==3.18.0

# Content Processing & Parsing
feedparser==6.0.11
html2text==2025.4.15
lxml==6.0.0
lxml_html_clean==0.4.2
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
mdurl==0.1.2
readability-lxml==0.8.4.1
soupsieve==2.7

# Web Scraping & Browser Automation
playwright==1.53.0
playwright-stealth==2.0.0
pyee==13.0.0

# YouTube Processing
pytube==15.0.0
youtube-transcript-api==1.1.1

# AI & LLM Integration
litellm==1.74.6
openai==1.97.0
huggingface-hub==0.33.4
tiktoken==0.9.0
tokenizers==0.21.2

# Testing Framework
pytest==8.4.1
pytest-cov==6.0.0
pytest-mock==3.14.1
responses==0.25.7

# Development Tools
black
isort
mypy
pre-commit==4.2.0
watchdog==6.0.0

# Data Processing & Utilities
frozenlist==1.7.0
fsspec==2025.7.0
h11==0.16.0
hf-xet==1.1.5
importlib_metadata==8.7.0
iniconfig==2.1.0
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
multidict==6.6.3
packaging==25.0
pluggy==1.6.0
propcache==0.3.2
Pygments==2.19.2
referencing==0.36.2
regex==2024.11.6
rich==14.0.0
rpds-py==0.26.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==5.3.1
yarl==1.20.1
zipp==3.23.0